refactor: pack LoanStorageData and externalize loan logic libraries
Pack 9 uint256 config params in LoanStorageData into address slots
(17 slots → 8 slots). Convert RepayLogic, CloseLoanLogic, and
FlashLoanLogic from internal to public linked libraries using the
_resolveStorage(bytes32) pattern established by LoanLogic.

Loan.sol: 24,806B → 16,016B (8,560B headroom).

Changes:
- src/protocol/LoanStorage.sol: packed struct layout
- src/interfaces/ILoan.sol: packed InitParams + setter types
- src/libraries/logic/{RepayLogic,CloseLoanLogic,FlashLoanLogic}.sol:
  internal → public with _resolveStorage(bytes32)
- src/protocol/Loan.sol: direct library calls with LOAN_STORAGE_LOCATION
- test/script: type cast updates for packed types
- script/deployment/DeployLibraries.s.sol: new script for library deployment
- script/deployment/DeploymentBase.s.sol: 3 new Phase1Addresses fields
- deploy/scripts/deploy-local.sh: use DeployLibraries.s.sol

+++ loan-provider/script/interaction/Loan.s.sol
@@ -71,7 +71,7 @@ contract Loan_SetGracePeriod is Script {
         ILoan loan = ILoan(loanAddress);
 
         vm.broadcast();
-        loan.setGracePeriod(gracePeriod);
+        loan.setGracePeriod(uint32(gracePeriod));
     }
 
     function _setGracePeriod() internal {
